requirements.txt
torch
torchvision
pandas
numpy
matplotlib
opencv-python
Pillow
timm
tqdm